Updates to support unit testing, data type review and proper serialization for R

+++ src/main/java/org/ohdsi/analysis/prediction/design/SeedSettings.java
@@ -2,7 +2,16 @@ package org.ohdsi.analysis.prediction.design;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
 
+/**
+ *
+ * @author Anthony Sena <https://github.com/anthonysena>
+ */
 public interface SeedSettings {
+
+    /**
+     *
+     * @return
+     */
     @JsonGetter("seed")
-    Float getSeed();
+    Integer getSeed();
 }
